# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SQLTunnel Gateway API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "SQLTunnelBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message", "requestId"],
        "properties": {
          "code": {
            "type": "string",
            "examples": ["QUERY_FAILED"]
          },
          "message": {
            "type": "string",
            "examples": ["Table 'app.users' doesn't exist"]
          },
          "requestId": {
            "type": "string",
            "examples": ["42d826dc-f2d5-4e6a-9267-1a378e6362d2"]
          }
        }
      },
      "QueryRequest": {
        "type": "object",
        "required": ["dbServerId", "sql"],
        "properties": {
          "dbServerId": {
            "type": "string"
          },
          "sql": {
            "type": "string"
          },
          "params": {
            "type": "array",
            "items": {}
          },
          "maxRows": {
            "type": "integer",
            "minimum": 1
          },
          "responseFormat": {
            "type": "string",
            "enum": ["raw", "json"],
            "default": "raw"
          }
        }
      },
      "QueryJsonResponse": {
        "type": "object",
        "required": ["columns", "rows", "rowCount", "durationMs", "dbServerId"],
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "rowCount": {
            "type": "integer",
            "minimum": 0
          },
          "durationMs": {
            "type": "integer",
            "minimum": 0
          },
          "dbServerId": {
            "type": "string"
          }
        }
      },
      "ListDatabasesSchemaRequest": {
        "type": "object",
        "required": ["operation"],
        "properties": {
          "operation": {
            "type": "string",
            "enum": ["list_databases"]
          }
        }
      },
      "ListTablesSchemaRequest": {
        "type": "object",
        "required": ["operation", "dbServerId"],
        "properties": {
          "operation": {
            "type": "string",
            "enum": ["list_tables"]
          },
          "dbServerId": {
            "type": "string"
          },
          "refresh": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "DescribeTableSchemaRequest": {
        "type": "object",
        "required": ["operation", "dbServerId", "schemaName", "tableName"],
        "properties": {
          "operation": {
            "type": "string",
            "enum": ["describe_table"]
          },
          "dbServerId": { "type": "string" },
          "schemaName": { "type": "string" },
          "tableName": { "type": "string" },
          "refresh": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "SchemaRequest": {
        "oneOf": [
          { "$ref": "#/components/schemas/ListDatabasesSchemaRequest" },
          { "$ref": "#/components/schemas/ListTablesSchemaRequest" },
          { "$ref": "#/components/schemas/DescribeTableSchemaRequest" }
        ],
        "discriminator": {
          "propertyName": "operation",
          "mapping": {
            "list_databases": "#/components/schemas/ListDatabasesSchemaRequest",
            "list_tables": "#/components/schemas/ListTablesSchemaRequest",
            "describe_table": "#/components/schemas/DescribeTableSchemaRequest"
          }
        }
      },
      "SchemaColumn": {
        "type": "object",
        "required": ["name", "dataType", "nullable", "ordinalPosition", "primaryKey", "unique"],
        "properties": {
          "name": { "type": "string" },
          "dataType": { "type": "string" },
          "nullable": { "type": "boolean" },
          "ordinalPosition": { "type": "integer", "minimum": 1 },
          "primaryKey": { "type": "boolean" },
          "unique": {
            "type": "boolean",
            "description": "Whether the column participates in a primary-key or unique constraint."
          },
          "defaultValue": { "type": "string" },
          "comment": { "type": "string" }
        }
      },
      "SchemaTable": {
        "type": "object",
        "required": ["name", "type", "columns"],
        "properties": {
          "name": { "type": "string" },
          "type": {
            "type": "string",
            "enum": ["table", "partitioned_table", "view", "materialized_view", "foreign_table"]
          },
          "comment": { "type": "string" },
          "columns": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SchemaColumn" }
          }
        }
      },
      "SchemaDatabaseSummary": {
        "type": "object",
        "required": ["dbServerId", "databaseName", "databaseType", "permission"],
        "properties": {
          "dbServerId": { "type": "string" },
          "databaseName": { "type": "string" },
          "databaseType": { "type": "string", "enum": ["mysql", "postgres"] },
          "permission": { "type": "string", "enum": ["read", "write"] }
        }
      },
      "SchemaTableSummary": {
        "type": "object",
        "required": ["schemaName", "tableName", "type"],
        "properties": {
          "schemaName": { "type": "string" },
          "tableName": { "type": "string" },
          "type": {
            "type": "string",
            "enum": ["table", "partitioned_table", "view", "materialized_view", "foreign_table"]
          },
          "comment": { "type": "string" }
        }
      },
      "ListDatabasesSchemaResponse": {
        "type": "object",
        "required": ["operation", "databases"],
        "properties": {
          "operation": { "type": "string", "enum": ["list_databases"] },
          "databases": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SchemaDatabaseSummary" }
          }
        }
      },
      "ListTablesSchemaResponse": {
        "type": "object",
        "required": ["operation", "dbServerId", "databaseName", "databaseType", "tables", "cached", "cachedAt"],
        "properties": {
          "operation": { "type": "string", "enum": ["list_tables"] },
          "dbServerId": { "type": "string" },
          "databaseName": { "type": "string" },
          "databaseType": { "type": "string", "enum": ["mysql", "postgres"] },
          "tables": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SchemaTableSummary" }
          },
          "cached": { "type": "boolean" },
          "cachedAt": { "type": "string", "format": "date-time" }
        }
      },
      "DescribeTableSchemaResponse": {
        "type": "object",
        "required": ["operation", "dbServerId", "databaseName", "databaseType", "schemaName", "table", "cached", "cachedAt"],
        "properties": {
          "operation": { "type": "string", "enum": ["describe_table"] },
          "dbServerId": { "type": "string" },
          "databaseName": { "type": "string" },
          "databaseType": { "type": "string", "enum": ["mysql", "postgres"] },
          "schemaName": { "type": "string" },
          "table": { "$ref": "#/components/schemas/SchemaTable" },
          "cached": { "type": "boolean" },
          "cachedAt": { "type": "string", "format": "date-time" }
        }
      },
      "SchemaResponse": {
        "oneOf": [
          { "$ref": "#/components/schemas/ListDatabasesSchemaResponse" },
          { "$ref": "#/components/schemas/ListTablesSchemaResponse" },
          { "$ref": "#/components/schemas/DescribeTableSchemaResponse" }
        ],
        "discriminator": {
          "propertyName": "operation"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request or SQL query failed. Common codes: INVALID_REQUEST, QUERY_FAILED, WRITE_FORBIDDEN.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthenticated": {
        "description": "Missing or invalid API key. Common code: UNAUTHENTICATED.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The API key cannot access the requested db server. Common code: FORBIDDEN.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "DbServerNotFound": {
        "description": "The requested db server was not found. Common code: DB_SERVER_NOT_FOUND.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "QueryTimeout": {
        "description": "The database query exceeded the configured query timeout. Common code: QUERY_TIMEOUT.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ConnectTimeout": {
        "description": "SSH tunnel setup or database connection exceeded the configured connection timeout. Common code: CONNECT_TIMEOUT.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal gateway or configuration error. Common codes: INTERNAL_ERROR, INVALID_CONFIG.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/schema": {
      "post": {
        "operationId": "schema",
        "summary": "List databases, list tables, or describe one table",
        "security": [
          {
            "SQLTunnelBearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Database metadata selected by the request operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/DbServerNotFound"
          },
          "408": {
            "$ref": "#/components/responses/QueryTimeout"
          },
          "413": {
            "description": "Schema exceeds the supported column limit. Common code: SCHEMA_TOO_LARGE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "$ref": "#/components/responses/ConnectTimeout"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/query": {
      "post": {
        "operationId": "query",
        "summary": "Execute one authorized SQL statement",
        "security": [
          {
            "SQLTunnelBearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query result. text/plain is returned by default; application/json is returned when responseFormat is json.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryJsonResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/DbServerNotFound"
          },
          "408": {
            "$ref": "#/components/responses/QueryTimeout"
          },
          "504": {
            "$ref": "#/components/responses/ConnectTimeout"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  }
}